+++ changelog.md
@@ -1,5 +1,12 @@
 # Range Slider Changelog
 
+## 1.0.0 - 2021-09-28
+### Fixed
+- Get it working with most recent Version of Craft CMS
+
+### Added
+- Added migration script to migrate data from `mx-rangeslider`
+
 ## 0.9.1 - 2019-01-08
 ### Fixed
 - Corrected wrong slider position when editing an existing entry. (Thanks @ushapax!)